// src/api-client/models/category-response-dto.ts
/* tslint:disable */

/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface CategoryResponseDto
 */
export interface CategoryResponseDto {
  /**
   *
   * @type {number}
   * @memberof CategoryResponseDto
   */
  id: number;
  /**
   *
   * @type {string}
   * @memberof CategoryResponseDto
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof CategoryResponseDto
   */
  code: string;
  /**
   *
   * @type {string}
   * @memberof CategoryResponseDto
   */
  description: string;
  /**
   * When the record was created
   * @type {string}
   * @memberof CategoryResponseDto
   */
  createdAt: string;
  /**
   * Soft-delete timestamp; null while the record is active
   * @type {string}
   * @memberof CategoryResponseDto
   */
  deletedAt: string | null;
}